+++ src/coguard_cli/discovery/config_file_finders/config_file_finder_terraform.py
@@ -107,4 +107,8 @@ class ConfigFileFinderTerraform(ConfigFileFinder):
         """
         return 'terraform'
 
+    def is_cluster_service(self):
+        return True
+
+
 ConfigFileFinder.register(ConfigFileFinderTerraform)
