+++ src/coguard_cli/docker_dao.py
@@ -249,28 +249,34 @@ def terraformer_wrapper(location_to_mount: str,
     """
     This function builds and runs the terraformer wrapper we built for CoGuard.
     """
-    # TODO: Make a test that fails in case we change the loc of this file.
     terraformer_wrapper_image_name = "terraformer_coguard_wrapper"
+    terraformer_container_name = "coguard_terraformer_extract"
     working_dir_of_this_file = os.path.dirname(os.path.abspath(__file__))
     try:
+        logging.info("Creating cloud extraction image on this machine.")
         subprocess.run(
             f"docker build -t {terraformer_wrapper_image_name} " + \
             os.path.join(working_dir_of_this_file,
                          "discovery",
                          "cloud_discovery",
-                         "terraformer_extract_image_helper"),
+                         "terraformer_extract_image_helper") + \
+            f" --build-arg USER_ID={os.getuid()} --build-arg GROUP_ID={os.getgid()}",
             check=True,
             shell=True,
             capture_output=True,
             timeout=DOCKER_CALL_TIMEOUT_S
         )
+        logging.info("Extracting your cloud configurations. This may take a while.")
         subprocess.run(
-            f"docker run --rm -v \"{location_to_mount}\":/opt/terraformer_export_data " + \
+            f"docker container stop {terraformer_container_name} || true && " + \
+            f"docker run --rm --name={terraformer_container_name} " + \
+            f"-v \"{location_to_mount}\":/opt/terraformer_export_data " + \
             " ".join(f"-e \"{k}\"=\"{v}\"" for k, v in environment_variables.items()) + \
             f" {terraformer_wrapper_image_name}",
             check=True,
             shell=True,
-            capture_output=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
             timeout=DOCKER_CALL_TIMEOUT_S * 10
         )
         return True
