+++ src/coguard_cli/tests/auth/common_auth_function_test.py
@@ -124,48 +124,6 @@ class TestAuthClass(unittest.TestCase):
             )
             self.assertIsNone(res)
 
-    def test_authenticate_to_server_empty_config_object(self):
-        """
-        Tests that the authentication returns None if the configuration object is None
-        """
-        self.assertIsNone(coguard_cli.auth.authenticate_to_server(None))
-
-    def test_authenticate_to_server_non_empty_config_object_404(self):
-        """
-        Tests that the authentication returns None if the status code
-        is not 200.
-        """
-        config_obj = coguard_cli.auth.auth_config.CoGuardCliConfig("foo", "bar")
-        def new_callable(url, data, timeout):
-            response = requests.Response()
-            setattr(response, 'status_code', 404)
-            return response
-        with unittest.mock.patch('requests.post',
-                                 new_callable=lambda: new_callable):
-            self.assertIsNone(coguard_cli.auth.authenticate_to_server(config_obj))
-
-    def test_authenticate_to_server_non_empty_config_object_success(self):
-        """
-        Tests that the authentication returns None if the status code
-        is not 200.
-        """
-        config_obj = coguard_cli.auth.auth_config.CoGuardCliConfig("foo", "bar")
-        def new_callable(url, data, timeout):
-            response = requests.Response()
-            setattr(response, 'status_code', 200)
-            return response
-        def new_json(arg):
-            return {
-                "access_token": "foo"
-            }
-        with unittest.mock.patch('requests.post',
-                                 new_callable=lambda: new_callable), \
-             unittest.mock.patch('requests.Response.json',
-                                 new_callable=lambda: new_json):
-            result = coguard_cli.auth.authenticate_to_server(config_obj)
-            self.assertIsNotNone(result)
-            self.assertEqual(result.get_token(False), "foo")
-
     def test_store_config_object_in_auth_file(self):
         """
         Testing the storing of the configuration object
@@ -186,178 +144,3 @@ class TestAuthClass(unittest.TestCase):
                 "foo/bar/baz.conf"
             )
             makedirs_mock.assert_called_with("foo/bar", exist_ok=True)
-
-    def test_get_public_key_status_not_200(self):
-        """
-        Tests the public key extraction of the authentication server.
-        """
-        mock_response = unittest.mock.Mock(
-            status_code = 420
-        )
-        mock_config_object = unittest.mock.Mock(
-            get_auth_url = lambda: "https://portal.coguard.io"
-        )
-        with unittest.mock.patch('requests.get',
-                                 new_callable=lambda: lambda url, timeout: mock_response):
-            result = coguard_cli.auth.get_public_key(
-                mock_config_object
-            )
-            self.assertIsNone(result)
-
-    def test_get_public_key_status_200_no_public_key(self):
-        """
-        Tests the public key extraction of the authentication server.
-        """
-        mock_response = unittest.mock.Mock(
-            status_code = 200,
-            json = lambda: {}
-        )
-        mock_config_object = unittest.mock.Mock(
-            get_auth_url = lambda: "https://portal.coguard.io/auth"
-        )
-        with unittest.mock.patch('requests.get',
-                                 new_callable=lambda: lambda url, timeout: mock_response):
-            result = coguard_cli.auth.get_public_key(
-                mock_config_object
-            )
-            self.assertIsNone(result)
-
-    def test_get_public_key_status_200_public_key(self):
-        """
-        Tests the public key extraction of the authentication server.
-        """
-        mock_response = unittest.mock.Mock(
-            status_code = 200,
-            json = lambda: {"public_key": "123456"}
-        )
-        mock_config_object = unittest.mock.Mock(
-            get_auth_url = lambda: "https://portal.coguard.io/auth"
-        )
-        with unittest.mock.patch('requests.get',
-                                 new_callable=lambda: lambda url, timeout: mock_response):
-            result = coguard_cli.auth.get_public_key(
-                mock_config_object
-            )
-            self.assertEqual(result, "123456")
-
-    def test_get_decoded_jwt_token(self):
-        """
-        Tests the decoded jwt token.
-        """
-        mock_result = unittest.mock.Mock()
-        mock_jwt = unittest.mock.Mock(
-            decode = lambda token, public_key: mock_result
-        )
-        with unittest.mock.patch(
-                'jwt.JWT',
-                return_value=mock_jwt
-        ), \
-        unittest.mock.patch(
-                'jwt.jwk_from_pem',
-                new_callable = lambda: lambda b: "foo"
-        ):
-            result = coguard_cli.auth.get_decoded_jwt_token(
-                unittest.mock.Mock(
-                    get_token = lambda x: "foo"),
-                "bar"
-            )
-            self.assertEqual(result, mock_result)
-
-    def test_extract_organization_from_token_no_pk(self):
-        """
-        Tests the extraction of the organization key from the token.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: ""
-        ):
-            result = coguard_cli.auth.extract_organization_from_token(
-                "foo", {}
-            )
-            self.assertIsNone(result)
-
-    def test_extract_organization_from_token_pk_empty(self):
-        """
-        Tests the extraction of the organization key from the token.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: "pk"
-        ), \
-        unittest.mock.patch(
-                'coguard_cli.auth.get_decoded_jwt_token',
-                new_callable=lambda: lambda token, public_key: {}
-        ):
-            result = coguard_cli.auth.extract_organization_from_token(
-                "foo", {}
-            )
-            self.assertIsNone(result)
-
-    def test_extract_organization_from_token_pk(self):
-        """
-        Tests the extraction of the organization key from the token.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: "pk"
-        ), \
-        unittest.mock.patch(
-                'coguard_cli.auth.get_decoded_jwt_token',
-                new_callable=lambda: lambda token, public_key: {"organization": "bar"}
-        ):
-            result = coguard_cli.auth.extract_organization_from_token(
-                "foo", {}
-            )
-            self.assertEqual(result, "bar")
-
-    def test_extract_deal_type_from_token_no_pk(self):
-        """
-        Tests the extraction of deal type from the token no public key.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: ""
-        ):
-            result = coguard_cli.auth.extract_deal_type_from_token(
-                "foo", {}
-            )
-            self.assertEqual(result, coguard_cli.auth.DealEnum.FREE)
-
-    def test_extract_deal_type_from_token_pk_no_deal_id(self):
-        """
-        Tests the extraction of deal type from the token no public key.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: "foo"
-        ), \
-        unittest.mock.patch(
-                'coguard_cli.auth.get_decoded_jwt_token',
-                new_callable=lambda: lambda token, pk: {}
-        ):
-            result = coguard_cli.auth.extract_deal_type_from_token(
-                "foo", {}
-            )
-            self.assertEqual(result, coguard_cli.auth.DealEnum.FREE)
-
-    def test_extract_deal_type_from_token_pk_deal_id(self):
-        """
-        Tests the extraction of deal type from the token no public key.
-        """
-        with unittest.mock.patch(
-                'coguard_cli.auth.get_public_key',
-                new_callable=lambda: lambda c: "foo"
-        ), \
-        unittest.mock.patch(
-                'coguard_cli.auth.get_decoded_jwt_token',
-                new_callable=lambda: lambda token, pk: {
-                    "realm_access": {
-                        "roles": [
-                            "DEV"
-                        ]
-                    }
-                }):
-            result = coguard_cli.auth.extract_deal_type_from_token(
-                "foo", {}
-            )
-            self.assertEqual(result, coguard_cli.auth.DealEnum.DEV)
