+++ src/coguard_cli/tests/image_check/common_functions_test.py
@@ -6,7 +6,7 @@ inside the image_check module.
 import unittest
 import unittest.mock
 from coguard_cli import image_check
-from coguard_cli.auth import DealEnum
+from coguard_cli.auth.util import DealEnum
 from coguard_cli.discovery.config_file_finders.config_file_finder_nginx \
     import ConfigFileFinderNginx
 from coguard_cli.discovery.config_file_finders.config_file_finder_netlify \
